+++ NEWS.md
@@ -1,6 +1,8 @@
 # bigrquery (development version)
 
-* `bq_table_stream()` - allows to stream data.frame into BigQuery table using [insertAll API endpoint](https://cloud.google.com/bigquery/docs/reference/rest/v2/tabledata/insertAll). 
+
+* `bq_perform_query()` can now execute parameterised query with parameters 
+  of ARRAY type (@byapparov, #303).
 
 # bigrquery 1.2.0
 
@@ -38,6 +40,8 @@ gargle and rlang are newly Imported.
 # bigrquery 1.1.1
 
 * Fix test failure with dbplyr 1.4.0.
+* `bq_perform_query()` can now execute parameterised query with parameters 
+  of ARRAY type (@byapparov, #303).
 
 * `bq_field()` can now pass `description` parameter which will be applied
   in `bq_table_create()` call (@byapparov, #272).